agentic-city-pulse-backend-server-main/agents/requirements.txt
# Core FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
google-adk
# Google AI and ADK dependencies
google-generativeai>=0.3.0
google-genai>=0.1.0

# Audio and speech processing
pydub>=0.25.1
SpeechRecognition>=3.10.0
pytube>=15.0.0

# Image processing
Pillow>=10.0.0
python-magic>=0.4.27

# HTTP and API requests
requests>=2.31.0
aiohttp>=3.9.0

# Reddit API
praw>=7.7.0

# Task scheduling
APScheduler>=3.10.0

# Data validation and serialization
pydantic>=2.5.0

# Environment and configuration
python-dotenv>=1.0.0

# Web framework for Streamlit app
streamlit>=1.28.0

# Additional utilities
typing-extensions>=4.8.0
